requirements.txt
substrate-interface>=1.7.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
requests>=2.31.0